// src/api/entities/tutanota/TypeRefs.ts
export type Birthday = {
	_type: TypeRef<Birthday>;

	_id: Id;
	day: NumberString;
	month: NumberString;
	year: null | NumberString;
}
// ...
export const ContactTypeRef: TypeRef<Contact> = new TypeRef("tutanota", "Contact")

export function createContact(values?: Partial<Contact>): Contact {
	return Object.assign(create(typeModels.Contact, ContactTypeRef), values)
}

export type Contact = {
	_type: TypeRef<Contact>;
	_errors: Object;

	_area: NumberString;
	_format: NumberString;
	_id: IdTuple;
	_owner: Id;
	_ownerEncSessionKey: null | Uint8Array;
	_ownerGroup: null | Id;
	_permissions: Id;
	autoTransmitPassword: string;
	birthdayIso: null | string;
	comment: string;
	company: string;
	firstName: string;
	lastName: string;
	nickname: null | string;
	oldBirthdayDate: null | Date;
	presharedPassword: null | string;
	role: string;
	title: null | string;

	addresses: ContactAddress[];
	mailAddresses: ContactMailAddress[];
	oldBirthdayAggregate:  null | Birthday;
	phoneNumbers: ContactPhoneNumber[];
	photo:  null | IdTuple;
	socialIds: ContactSocialId[];
}
export const ContactAddressTypeRef: TypeRef<ContactAddress> = new TypeRef("tutanota", "ContactAddress")

export function createContactAddress(values?: Partial<ContactAddress>): ContactAddress {
	return Object.assign(create(typeModels.ContactAddress, ContactAddressTypeRef), values)
}

export type ContactAddress = {
	_type: TypeRef<ContactAddress>;

	_id: Id;
	address: string;
	customTypeName: string;
	type: NumberString;
}
// ...
export type ContactMailAddress = {
	_type: TypeRef<ContactMailAddress>;

	_id: Id;
	address: string;
	customTypeName: string;
	type: NumberString;
}
export const ContactPhoneNumberTypeRef: TypeRef<ContactPhoneNumber> = new TypeRef("tutanota", "ContactPhoneNumber")

export function createContactPhoneNumber(values?: Partial<ContactPhoneNumber>): ContactPhoneNumber {
	return Object.assign(create(typeModels.ContactPhoneNumber, ContactPhoneNumberTypeRef), values)
}

export type ContactPhoneNumber = {
	_type: TypeRef<ContactPhoneNumber>;

	_id: Id;
	customTypeName: string;
	number: string;
	type: NumberString;
}
export const ContactSocialIdTypeRef: TypeRef<ContactSocialId> = new TypeRef("tutanota", "ContactSocialId")

export function createContactSocialId(values?: Partial<ContactSocialId>): ContactSocialId {
	return Object.assign(create(typeModels.ContactSocialId, ContactSocialIdTypeRef), values)
}

export type ContactSocialId = {
	_type: TypeRef<ContactSocialId>;

	_id: Id;
	customTypeName: string;
	socialId: string;
	type: NumberString;
}
// ...
export type EncryptedMailAddress = {
	_type: TypeRef<EncryptedMailAddress>;

	_id: Id;
	address: string;
	name: string;
}
// ...
export const MailTypeRef: TypeRef<Mail> = new TypeRef("tutanota", "Mail")

export function createMail(values?: Partial<Mail>): Mail {
	return Object.assign(create(typeModels.Mail, MailTypeRef), values)
}

export type Mail = {
	_type: TypeRef<Mail>;
	_errors: Object;

	_format: NumberString;
	_id: IdTuple;
	_ownerEncSessionKey: null | Uint8Array;
	_ownerGroup: null | Id;
	_permissions: Id;
	authStatus: null | NumberString;
	confidential: boolean;
	differentEnvelopeSender: null | string;
	listUnsubscribe: boolean;
	method: NumberString;
	movedTime: null | Date;
	phishingStatus: NumberString;
	receivedDate: Date;
	recipientCount: NumberString;
	replyType: NumberString;
	sentDate: null | Date;
	state: NumberString;
	subject: string;
	unread: boolean;

	attachments: IdTuple[];
	bccRecipients: MailAddress[];
	body:  null | Id;
	bucketKey:  null | BucketKey;
	ccRecipients: MailAddress[];
	conversationEntry: IdTuple;
	firstRecipient:  null | MailAddress;
	headers:  null | Id;
	mailDetails:  null | IdTuple;
	mailDetailsDraft:  null | IdTuple;
	replyTos: EncryptedMailAddress[];
	restrictions:  null | MailRestriction;
	sender: MailAddress;
	toRecipients: MailAddress[];
}
export const MailAddressTypeRef: TypeRef<MailAddress> = new TypeRef("tutanota", "MailAddress")

export function createMailAddress(values?: Partial<MailAddress>): MailAddress {
	return Object.assign(create(typeModels.MailAddress, MailAddressTypeRef), values)
}

export type MailAddress = {
	_type: TypeRef<MailAddress>;

	_id: Id;
	address: string;
	name: string;

	contact:  null | IdTuple;
}
// ...
export type MailRestriction = {
	_type: TypeRef<MailRestriction>;

	_id: Id;

	delegationGroups_removed: Id[];
	participantGroupInfos: IdTuple[];
}